Move SchemaHandler and dependencies into wire-schema's commonMain
Also inlined wire-profiles into wire-schema to avoid circular deps
