+++ settings.gradle.kts
@@ -33,7 +33,6 @@ includeBuild("wire-library") {
     substitute(module("com.squareup.wire:wire-java-generator")).with(project(":wire-java-generator"))
     substitute(module("com.squareup.wire:wire-kotlin-generator")).with(project(":wire-kotlin-generator"))
     substitute(module("com.squareup.wire:wire-moshi-adapter")).with(project(":wire-moshi-adapter"))
-    substitute(module("com.squareup.wire:wire-profiles")).with(project(":wire-profiles"))
     substitute(module("com.squareup.wire:wire-runtime")).with(project(":wire-runtime"))
     substitute(module("com.squareup.wire:wire-schema")).with(project(":wire-schema"))
     substitute(module("com.squareup.wire:wire-schema-tests")).with(project(":wire-schema-tests"))
