+++ wire-library/settings.gradle.kts
@@ -9,7 +9,6 @@ include(":wire-gson-support")
 include(":wire-java-generator")
 include(":wire-kotlin-generator")
 include(":wire-moshi-adapter")
-include(":wire-profiles")
 include(":wire-reflector")
 include(":wire-runtime:japicmp")
 include(":wire-runtime")
