+++ wire-library/wire-compiler/build.gradle.kts
@@ -21,7 +21,6 @@ dependencies {
   implementation(project(":wire-kotlin-generator"))
   implementation(project(":wire-java-generator"))
   implementation(project(":wire-swift-generator"))
-  implementation(project(":wire-profiles"))
   implementation(deps.okio.core)
   implementation(deps.guava)
   implementation(deps.kotlin.serialization)
