+++ wire-library/wire-grpc-server-generator/build.gradle.kts
@@ -12,7 +12,6 @@ plugins {
 
 dependencies {
   api(project(":wire-schema"))
-  implementation(project(":wire-profiles"))
   implementation(project(":wire-runtime"))
   implementation(project(":wire-grpc-client"))
   implementation(deps.okio.core)
