+++ wire-library/wire-java-generator/build.gradle.kts
@@ -12,7 +12,6 @@ plugins {
 dependencies {
   api(project(":wire-schema"))
   implementation(project(":wire-runtime"))
-  implementation(project(":wire-profiles"))
   implementation(deps.okio.core)
   implementation(deps.guava)
   api(deps.javapoet)
