--- wire-library/wire-profiles/build.gradle.kts
@@ -1,31 +1,0 @@
-import com.vanniktech.maven.publish.JavadocJar.Dokka
-import com.vanniktech.maven.publish.KotlinJvm
-import com.vanniktech.maven.publish.MavenPublishBaseExtension
-
-plugins {
-  id("java-library")
-  kotlin("jvm")
-  id("org.jetbrains.dokka")
-  id("com.vanniktech.maven.publish.base")
-}
-
-dependencies {
-  api(project(":wire-schema"))
-  implementation(project(":wire-runtime"))
-  implementation(deps.okio.core)
-  implementation(deps.guava)
-  api(deps.javapoet)
-  api(deps.kotlinpoet)
-  compileOnly(deps.jsr305)
-  testImplementation(project(":wire-test-utils"))
-  testImplementation(deps.junit)
-  testImplementation(deps.kotlin.test.junit)
-  testImplementation(deps.assertj)
-  testImplementation(deps.jimfs)
-}
-
-configure<MavenPublishBaseExtension> {
-  configure(
-    KotlinJvm(javadocJar = Dokka("dokkaGfm"), sourcesJar = true)
-  )
-}
