+++ wire-library/wire-schema-tests/build.gradle.kts
@@ -15,7 +15,6 @@ dependencies {
   implementation(project(":wire-kotlin-generator"))
   implementation(project(":wire-java-generator"))
   implementation(project(":wire-swift-generator"))
-  implementation(project(":wire-profiles"))
   implementation(deps.okio.core)
   implementation(deps.okio.fakefilesystem)
   testImplementation(deps.assertj)
