+++ wire-library/wire-schema/src/commonMain/kotlin/com/squareup/wire/WireLogger.kt
