+++ wire-library/wire-schema/src/commonMain/kotlin/com/squareup/wire/schema/ClaimedDefinitions.kt
