+++ wire-library/wire-schema/src/commonMain/kotlin/com/squareup/wire/schema/ClaimedPaths.kt
@@ -27,22 +27,26 @@ class ClaimedPaths {
 
   /** Tracks that [type] has been generated to [path]. */
   fun claim(path: Path, type: Type) {
-    val existingEntry = paths.putIfAbsent(path, type.asErrorMessage())
+    val errorMessage = type.asErrorMessage()
+    val existingEntry = paths[path]
     check(existingEntry == null) {
       "Same file $path is getting generated by different messages:\n" +
         "  ${existingEntry}\n" +
         "  ${type.asErrorMessage()}"
     }
+    paths[path] = errorMessage
   }
 
   /** Tracks that [service] has been generated to [path]. */
   fun claim(path: Path, service: Service) {
-    val existingEntry = paths.putIfAbsent(path, service.asErrorMessage())
+    val errorMessage = service.asErrorMessage()
+    val existingEntry = paths[path]
     check(existingEntry == null) {
       "Same file $path is getting generated by different services:\n" +
         "  ${existingEntry}\n" +
         "  ${service.asErrorMessage()}"
     }
+    paths[path] = errorMessage
   }
 
   private fun Type.asErrorMessage(): String = "${type.simpleName} at $location"
