+++ wire-library/wire-schema/src/commonMain/kotlin/com/squareup/wire/schema/EmittingRules.kt
