+++ wire-library/wire-schema/src/commonMain/kotlin/com/squareup/wire/schema/Profile.kt
@@ -1,0 +1,3 @@
+package com.squareup.wire.schema
+
+expect class Profile
