+++ wire-library/wire-schema/src/jsMain/kotlin/com/squareup/wire/schema/Profile.kt
@@ -1,0 +1,3 @@
+package com.squareup.wire.schema
+
+actual class Profile
